refactor profile to MobileUserProfileController
