+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/mobile/MobileUserController.scala
@@ -304,33 +304,6 @@ class MobileUserController @Inject() (
     Ok
   }
 
-  def profile(username: String) = MaybeUserAction { request =>
-    val viewer = request.userOpt
-    userCommander.profile(Username(username), viewer) match {
-      case None => NotFound(s"can't find username $username")
-      case Some(profile) =>
-        val (numLibraries, numFollowedLibs, numInvitedLibs) = libraryCommander.countLibraries(profile.userId, viewer.map(_.id.get))
-        val (numConnections, userBiography) = db.readOnlyMaster { implicit s =>
-          val numConnections = userConnectionRepo.getConnectionCount(profile.userId)
-          val userBio = userValueRepo.getValueStringOpt(profile.userId, UserValueName.USER_DESCRIPTION)
-          (numConnections, userBio)
-        }
-
-        val jsonFriendInfo = Json.toJson(profile.basicUserWithFriendStatus).as[JsObject]
-        val jsonProfileInfo = Json.toJson(UserProfileStats(
-          numLibraries = numLibraries,
-          numFollowedLibraries = numFollowedLibs,
-          numKeeps = profile.numKeeps,
-          numConnections = numConnections,
-          numFollowers = libraryCommander.countFollowers(profile.userId, viewer.map(_.id.get)),
-          numInvitedLibraries = numInvitedLibs,
-          biography = userBiography
-        )).as[JsObject]
-
-        Ok(jsonFriendInfo ++ jsonProfileInfo)
-    }
-  }
-
   def uploadBinaryUserPicture() = UserAction(parse.maxLength(1024 * 1024 * 15, parse.temporaryFile)) { implicit request =>
     request.body match {
       case Right(tempFile) =>
