+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/mobile/MobileUserProfileController.scala
@@ -22,6 +22,7 @@ class MobileUserProfileController @Inject() (
   basicUserRepo: BasicUserRepo,
   userConnectionRepo: UserConnectionRepo,
   libraryRepo: LibraryRepo,
+  userValueRepo: UserValueRepo,
   userCommander: UserCommander,
   userProfileCommander: UserProfileCommander,
   userConnectionsCommander: UserConnectionsCommander,
@@ -29,6 +30,33 @@ class MobileUserProfileController @Inject() (
   libraryCommander: LibraryCommander,
   implicit val executionContext: ExecutionContext)
     extends UserActions with ShoeboxServiceController {
+
+  def profile(username: String) = MaybeUserAction { request =>
+    val viewer = request.userOpt
+    userCommander.profile(Username(username), viewer) match {
+      case None => NotFound(s"can't find username $username")
+      case Some(profile) =>
+        val (numLibraries, numFollowedLibs, numInvitedLibs) = libraryCommander.countLibraries(profile.userId, viewer.map(_.id.get))
+        val (numConnections, userBiography) = db.readOnlyMaster { implicit s =>
+          val numConnections = userConnectionRepo.getConnectionCount(profile.userId)
+          val userBio = userValueRepo.getValueStringOpt(profile.userId, UserValueName.USER_DESCRIPTION)
+          (numConnections, userBio)
+        }
+
+        val jsonFriendInfo = Json.toJson(profile.basicUserWithFriendStatus).as[JsObject]
+        val jsonProfileInfo = Json.toJson(UserProfileStats(
+          numLibraries = numLibraries,
+          numFollowedLibraries = numFollowedLibs,
+          numKeeps = profile.numKeeps,
+          numConnections = numConnections,
+          numFollowers = libraryCommander.countFollowers(profile.userId, viewer.map(_.id.get)),
+          numInvitedLibraries = numInvitedLibs,
+          biography = userBiography
+        )).as[JsObject]
+
+        Ok(jsonFriendInfo ++ jsonProfileInfo)
+    }
+  }
 
   def getProfileFollowers(username: Username, page: Int = 0, size: Int = 12) = MaybeUserAction.async { request =>
     userCommander.userFromUsername(username) match {
