+++ kifi-backend/modules/shoebox/conf/shoeboxService.routes
@@ -227,7 +227,7 @@ GET     /m/1/contacts/search                    @com.keepit.controllers.mobile.M
 POST    /m/1/invite                             @com.keepit.controllers.mobile.MobileInviteController.inviteConnection
 
 # Profile Page
-GET     /m/1/user/:username/profile             @com.keepit.controllers.mobile.MobileUserController.profile(username)
+GET     /m/1/user/:username/profile             @com.keepit.controllers.mobile.MobileUserProfileController.profile(username)
 GET     /m/1/user/:username/libraries           @com.keepit.controllers.mobile.MobileLibraryController.getProfileLibraries(username: Username, page: Int ?= 0, size: Int ?= 12, filter: String ?= "own")
 GET     /m/1/user/:username/followers           @com.keepit.controllers.mobile.MobileUserProfileController.getProfileFollowers(username: Username, page: Int ?= 0, size: Int ?= 12)
